requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
sqlalchemy
pymysql
passlib[bcrypt]
email-validator
cryptography
python-dateutil
finance-datareader
pandas
lxml
openai>=1.0.0